+++ benches/benches/generators.rs
@@ -40,15 +40,15 @@ pub fn random_bytes(c: &mut Criterion) {
     }
 
     bench(&mut g, "step", StepRng::new(0, 1));
-    bench(&mut g, "pcg32", Pcg32::from_os_rng());
-    bench(&mut g, "pcg64", Pcg64::from_os_rng());
-    bench(&mut g, "pcg64mcg", Pcg64Mcg::from_os_rng());
-    bench(&mut g, "pcg64dxsm", Pcg64Dxsm::from_os_rng());
-    bench(&mut g, "chacha8", ChaCha8Rng::from_os_rng());
-    bench(&mut g, "chacha12", ChaCha12Rng::from_os_rng());
-    bench(&mut g, "chacha20", ChaCha20Rng::from_os_rng());
-    bench(&mut g, "std", StdRng::from_os_rng());
-    bench(&mut g, "small", SmallRng::from_thread_rng());
+    bench(&mut g, "pcg32", Pcg32::from_rng(&mut rand::rng()));
+    bench(&mut g, "pcg64", Pcg64::from_rng(&mut rand::rng()));
+    bench(&mut g, "pcg64mcg", Pcg64Mcg::from_rng(&mut rand::rng()));
+    bench(&mut g, "pcg64dxsm", Pcg64Dxsm::from_rng(&mut rand::rng()));
+    bench(&mut g, "chacha8", ChaCha8Rng::from_rng(&mut rand::rng()));
+    bench(&mut g, "chacha12", ChaCha12Rng::from_rng(&mut rand::rng()));
+    bench(&mut g, "chacha20", ChaCha20Rng::from_rng(&mut rand::rng()));
+    bench(&mut g, "std", StdRng::from_rng(&mut rand::rng()));
+    bench(&mut g, "small", SmallRng::from_rng(&mut rand::rng()));
     bench(&mut g, "os", UnwrapErr(OsRng));
     bench(&mut g, "thread", rand::rng());
 
@@ -69,15 +69,15 @@ pub fn random_u32(c: &mut Criterion) {
     }
 
     bench(&mut g, "step", StepRng::new(0, 1));
-    bench(&mut g, "pcg32", Pcg32::from_os_rng());
-    bench(&mut g, "pcg64", Pcg64::from_os_rng());
-    bench(&mut g, "pcg64mcg", Pcg64Mcg::from_os_rng());
-    bench(&mut g, "pcg64dxsm", Pcg64Dxsm::from_os_rng());
-    bench(&mut g, "chacha8", ChaCha8Rng::from_os_rng());
-    bench(&mut g, "chacha12", ChaCha12Rng::from_os_rng());
-    bench(&mut g, "chacha20", ChaCha20Rng::from_os_rng());
-    bench(&mut g, "std", StdRng::from_os_rng());
-    bench(&mut g, "small", SmallRng::from_thread_rng());
+    bench(&mut g, "pcg32", Pcg32::from_rng(&mut rand::rng()));
+    bench(&mut g, "pcg64", Pcg64::from_rng(&mut rand::rng()));
+    bench(&mut g, "pcg64mcg", Pcg64Mcg::from_rng(&mut rand::rng()));
+    bench(&mut g, "pcg64dxsm", Pcg64Dxsm::from_rng(&mut rand::rng()));
+    bench(&mut g, "chacha8", ChaCha8Rng::from_rng(&mut rand::rng()));
+    bench(&mut g, "chacha12", ChaCha12Rng::from_rng(&mut rand::rng()));
+    bench(&mut g, "chacha20", ChaCha20Rng::from_rng(&mut rand::rng()));
+    bench(&mut g, "std", StdRng::from_rng(&mut rand::rng()));
+    bench(&mut g, "small", SmallRng::from_rng(&mut rand::rng()));
     bench(&mut g, "os", UnwrapErr(OsRng));
     bench(&mut g, "thread", rand::rng());
 
@@ -98,15 +98,15 @@ pub fn random_u64(c: &mut Criterion) {
     }
 
     bench(&mut g, "step", StepRng::new(0, 1));
-    bench(&mut g, "pcg32", Pcg32::from_os_rng());
-    bench(&mut g, "pcg64", Pcg64::from_os_rng());
-    bench(&mut g, "pcg64mcg", Pcg64Mcg::from_os_rng());
-    bench(&mut g, "pcg64dxsm", Pcg64Dxsm::from_os_rng());
-    bench(&mut g, "chacha8", ChaCha8Rng::from_os_rng());
-    bench(&mut g, "chacha12", ChaCha12Rng::from_os_rng());
-    bench(&mut g, "chacha20", ChaCha20Rng::from_os_rng());
-    bench(&mut g, "std", StdRng::from_os_rng());
-    bench(&mut g, "small", SmallRng::from_thread_rng());
+    bench(&mut g, "pcg32", Pcg32::from_rng(&mut rand::rng()));
+    bench(&mut g, "pcg64", Pcg64::from_rng(&mut rand::rng()));
+    bench(&mut g, "pcg64mcg", Pcg64Mcg::from_rng(&mut rand::rng()));
+    bench(&mut g, "pcg64dxsm", Pcg64Dxsm::from_rng(&mut rand::rng()));
+    bench(&mut g, "chacha8", ChaCha8Rng::from_rng(&mut rand::rng()));
+    bench(&mut g, "chacha12", ChaCha12Rng::from_rng(&mut rand::rng()));
+    bench(&mut g, "chacha20", ChaCha20Rng::from_rng(&mut rand::rng()));
+    bench(&mut g, "std", StdRng::from_rng(&mut rand::rng()));
+    bench(&mut g, "small", SmallRng::from_rng(&mut rand::rng()));
     bench(&mut g, "os", UnwrapErr(OsRng));
     bench(&mut g, "thread", rand::rng());
 
@@ -120,7 +120,7 @@ pub fn init_gen(c: &mut Criterion) {
 
     fn bench<R: SeedableRng>(g: &mut BenchmarkGroup<WallTime>, name: &str) {
         g.bench_function(name, |b| {
-            let mut rng = Pcg32::from_os_rng();
+            let mut rng = Pcg32::from_rng(&mut rand::rng());
             b.iter(|| R::from_rng(&mut rng));
         });
     }
@@ -145,7 +145,7 @@ pub fn init_from_u64(c: &mut Criterion) {
 
     fn bench<R: SeedableRng>(g: &mut BenchmarkGroup<WallTime>, name: &str) {
         g.bench_function(name, |b| {
-            let mut rng = Pcg32::from_os_rng();
+            let mut rng = Pcg32::from_rng(&mut rand::rng());
             let seed = rng.random();
             b.iter(|| R::seed_from_u64(black_box(seed)));
         });
@@ -174,7 +174,7 @@ pub fn init_from_seed(c: &mut Criterion) {
         rand::distr::StandardUniform: Distribution<<R as SeedableRng>::Seed>,
     {
         g.bench_function(name, |b| {
-            let mut rng = Pcg32::from_os_rng();
+            let mut rng = Pcg32::from_rng(&mut rand::rng());
             let seed = rng.random();
             b.iter(|| R::from_seed(black_box(seed.clone())));
         });